Update Echopay docs and config; remove CSV
Refresh Echopay documentation and configuration: reorganise README (new Echopay NX header, password hint, AI prompt snippet, Typical rates and Colours sections) and remove legacy markdown/usages. Update config.json (description -> "Orders™", defaultTheme -> "light"). Remove large Soho_Vapes.csv and obsolete echopay markdown/types files. Also include minor updates to flash components, manifest and markdown index to align with the docs refresh.
