+++ public/echopay/config.json
@@ -1,7 +1,7 @@
 {
     "tenant": "echopay",
     "siteName": "EchoPay",
-    "description": "Open Banking",
+    "description": "Orders™",
     "url": "https://echopay.goldlabel.pro",
     "owner": {
         "name": "James Ward",
@@ -29,7 +29,7 @@
         },
         "designSystem": {
             "themeSwitching": true,
-            "defaultTheme": "dark",
+            "defaultTheme": "light",
             "themes": {
                 "dark": {
                     "mode": "dark",
