+++ public/echopay/markdown/index.md
@@ -2,24 +2,17 @@
 order: 1
 slug: /
 title: EchoPay
-description: Open Banking
-icon: shop
-tags: EchoPay, Flash, API, NX, Payments, Open Banking,
-flashXXX: EchoPay
+description: EchoPay Orders™ Summer 2026
+tags: B2B ordering app, EchoPay Orders, wholesale shopping, PWA ordering, fast product search, secure checkout,
+image: https://live.staticflickr.com/65535/55158091848_c9c249983f_b.jpg
 ---
 
-> "This new technology is also behind several tools that promote financial literacy and a greater control over one’s personal finances."
+> Shop smarter and faster with EchoPay Orders™. Search products, order anytime, and pay securely with EchoPay. EchoPay Orders™ puts the full catalogue in the palm of your hand.
 
-Open banking is still in its early years, but its growth is clear. However, there are still plenty of misconceptions about open banking and one of those is its impact on your credit score.
+- Browse and search instantly.
+- Order anytime, anywhere on any device.
+- Pay securely with EchoPay.
+- Stay in control with live stock, promotions, and order management.
 
-#### Future proofed payment
-
-Ultimately, open banking does not have a direct impact on your credit score. This is because open banking allows you to securely share your data with third-party suppliers, like lenders and financial service providers. 
-
-According to GoCardless, open banking can help reduce the stress associated with creditworthiness, as it keeps the process private, speeds up the application , and offers an improved customer experience. 
-
-The report said: “Open banking credit scoring solutions provide lenders with valuable data that allows them to make better decisions, levelling the competitive playing field. 
-
-#### Pay-by-Link 
-
-Pay-by-Link has been made even more straightforward with our new card tokenisation process. Stored encrypted in the cloud, our tokenisation allows for the customer card details to be saved on your account, enabling seamless 2-click payments. No more reading out card numbers on the phone!
+No queues. No paper forms. Just fast, convenient shopping.
+Sign up for early access and be the first to try
